:recycle: [Refactor] 어노테이션 클래스, global 위치로 이동
