+++ src/main/java/samryong/domain/item/contorller/ItemController.java
@@ -16,8 +16,8 @@ import samryong.domain.item.service.category.CategoryService;
 import samryong.domain.item.service.item.ItemService;
 import samryong.domain.location.dto.LocationDTO.LocationRequestDTO;
 import samryong.domain.member.entity.Member;
+import samryong.global.annotation.AuthMember;
 import samryong.global.response.ApiResponse;
-import samryong.security.resolver.annotation.AuthMember;
 
 @RestController
 @RequiredArgsConstructor
