+++ src/main/java/samryong/global/annotation/AuthChatMember.java
@@ -1,10 +1,11 @@
-package samryong.domain.chat.aspect.annotation;
+package samryong.global.annotation;
 
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
 import java.lang.annotation.RetentionPolicy;
 import java.lang.annotation.Target;
 
+// 해당 사용자가 채팅방에 속해있는지 확인하는 어노테이션
 @Retention(RetentionPolicy.RUNTIME)
 @Target(ElementType.METHOD)
 public @interface AuthChatMember {}
