+++ src/main/java/samryong/global/annotation/GetToken.java
@@ -1,10 +1,11 @@
-package samryong.security.resolver.annotation;
+package samryong.global.annotation;
 
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
 import java.lang.annotation.RetentionPolicy;
 import java.lang.annotation.Target;
 
+// 해당 사용자의 토큰을 가져오는 어노테이션
 @Target(ElementType.PARAMETER)
 @Retention(RetentionPolicy.RUNTIME)
 public @interface GetToken {}
